refactor: PointsGenerator 패키지 생성
